Set a data type for MDP transition probability arrays

+++ MDP_EM/MDP_EM/inference_mdp.py
@@ -19,7 +19,7 @@ class InferenceMDP(MDP):
     implemented as a dictionary: state: a subset of AP."""
     def __init__(self, init=None, action_list=[], states=[], prob=dict([]), gamma=.9, AP=set([]), L=dict([]),
                  reward=dict([]), grid_map=None, act_prob=dict([]), gg_kernel_centers=frozenset([]),
-                 og_kernel_centers=frozenset([]), kernel_sigmas=None):
+                 og_kernel_centers=frozenset([]), kernel_sigmas=None, prob_dtype=np.float64):
         """
         @brief Construct an MDP meant to perform inference.
         @param init @todo
@@ -38,7 +38,8 @@ class InferenceMDP(MDP):
         @param kernel_sigmas a @ref numpy.array() of length G+O of standard deviations.
         """
         super(self.__class__, self).__init__(init=init, action_list=action_list, states=states, prob=prob, gamma=gamma,
-                                             AP=AP, L=L, reward=reward, grid_map=grid_map, act_prob=act_prob)
+                                             AP=AP, L=L, reward=reward, grid_map=grid_map, act_prob=act_prob,
+                                             prob_dtype=prob_dtype)
 
         self.graph = GridGraph(grid_map=grid_map, neighbor_dict=self.neighbor_dict, label_dict=self.L)
